% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaUtils.R
\name{calc_Nx_Lx}
\alias{calc_Nx_Lx}
\title{Calculate assembly statistics}
\usage{
calc_Nx_Lx(contig_sizes, x)
}
\arguments{
\item{contig_sizes}{a numeric vector of contig sizes}

\item{x}{a numeric value (most common are 50 and 90)}
}
\value{
cat() Nx and Lx values
}
\description{
Calculate assembly statistics
}
