% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FastaUtils_functions.R
\name{DNA2RNA}
\alias{DNA2RNA}
\title{Convert a DNA fasta to an RNA fasta}
\usage{
DNA2RNA(file = NULL, out = paste(file, ".rna", sep = ""))
}
\arguments{
\item{file}{Fasta file}

\item{out}{Path to output file. If absent, the '.rna' is added to the input file's name (path is thus conserved).}
}
\value{
Writes a fasta file.
}
\description{
This function loads a fasta filecontaining DNA sequences and writes a fasta file with the same sequences converted to its RNA equivalent.
}
\examples{
DNA2RNA(file="http://greengenes.lbl.gov/Data/JD_Tutorial/UnAligSeq24606.txt",out="out.fasta")
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{FastaUtils}

