% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaUtils_functions.R
\name{fasta.sample}
\alias{fasta.sample}
\title{Select a random sample of sequences}
\usage{
fasta.sample(infile = NULL, nseq = NULL, file.out = NULL, replacement = FALSE)
}
\arguments{
\item{infile}{Multi-Fasta file}

\item{nseq}{Sample size.}

\item{file.out}{Path to output file.}

\item{replacement}{Logical for using or not (default) sampling with replacement.}
}
\value{
Writes a fasta file with sampled sequences.
}
\description{
This function select a random sample of sequences with or without replacement.
}
\examples{
fasta.sample(infile="http://greengenes.lbl.gov/Data/JD_Tutorial/UnAligSeq24606.txt",nseq=10,file.out="out.fasta")
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{FastaUtils}
