% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastagrab.R
\name{fastagrab}
\alias{fastagrab}
\title{fastagrab() extracts a fasta sequence from a multi fasta file}
\usage{
fastagrab(fasta, name, out_fasta)
}
\arguments{
\item{fasta}{path of the fasta file}

\item{name}{name of the sequence to extract (with our without '>')}

\item{out_fasta}{name and full path of the file in
which to write the extracted sequence}
}
\value{
nothing
}
\description{
description
}
