% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaUtils.R
\name{fastaList}
\alias{fastaList}
\title{Lists contigs of a (multi) fasta file and
counts contig sizes}
\usage{
fastaList(fasta, metrics = "none")
}
\arguments{
\item{fasta}{path of the fasta file}

\item{metrics}{a character vector of metrics to return in c("N50") default:
"none"}
}
\value{
number of contigs and contigs sizes, possibly some metrics
}
\description{
Lists contigs of a (multi) fasta file and
counts contig sizes
}
