% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastanalyze.R
\name{fastanalyze}
\alias{fastanalyze}
\title{fastanalyze() gives information about a fasta file}
\usage{
fastanalyze(fasta_file, metrics = TRUE, plot = FALSE, verbose = TRUE)
}
\arguments{
\item{fasta_file}{path of the fasta file}

\item{metrics}{boolean, indicates whether the metrics should be printed}

\item{plot}{boolean, indicates whether a plot should be generated}

\item{verbose}{boolean, indicates whether each contig name and size be
reported. default: TRUE}
}
\value{
Whatever you want
}
\description{
it counts the number of contigs, their size and the total size.
  It also calculates the N50 / L50, the N90 / L90 and prints a synthetic plot
  suming up all these results
}
