% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FastaUtils_functions.R
\docType{package}
\name{FastaUtils}
\alias{FastaUtils}
\alias{FastaUtils-package}
\title{FastaUtils: Utilities for DNA/RNA sequence processing}
\description{
The package \pkg{FastaUtils} provides tools for the manipulation of DNA/RNA sequence data.
}
\details{
The package \pkg{FastaUtils} depends on \pkg{Biostrings} which can be installed from Bioconductor at http://bioconductor.org/packages/release/bioc/html/Biostrings.html

To see the preferable citation of the package, type citation("FastaUtils").
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}

