% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx_lx.R
\name{nx_lx}
\alias{nx_lx}
\title{Calculate assembly statistics N50, L50, N90, L90}
\usage{
nx_lx(contig_df)
}
\arguments{
\item{contigs_df, }{a contigs data.frame with two cols, name and size}
}
\value{
a data.frame containing N50, L50, N90, L90 and
the input data.frame with cols cum_size and contig_n added
}
\description{
Calculate assembly statistics N50, L50, N90, L90
}
