% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{List contigs names and sizes from a fasta file.}
\usage{
read_fasta(fasta)
}
\arguments{
\item{fasta}{path of the fasta file}
}
\value{
a data.frame with two cols, name & size
}
\description{
List contigs names and sizes from a fasta file.
}
