% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaUtils_functions.R
\name{RNA2DNA}
\alias{RNA2DNA}
\title{Convert a RNA fasta to an DNA fasta}
\usage{
RNA2DNA(file = NULL, out = paste(file, ".dna", sep = ""))
}
\arguments{
\item{file}{Fasta file}

\item{out}{Path to output file. If absent, the '.dna' is added to the input file's name (path is thus conserved).}
}
\value{
Writes a fasta file.
}
\description{
This function loads a fasta filecontaining RNA sequences and writes a fasta file with the same sequences converted to its DNA equivalent.
}
\examples{
# See 'DNA2RNA()' function
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{FastaUtils}
