% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FastaUtils_functions.R
\name{subset.fasta}
\alias{subset.fasta}
\title{Select a subset of sequences from a fasta file}
\usage{
\method{subset}{fasta}(file = NULL, subset = NULL, out = paste(file,
  ".subset", sep = ""))
}
\arguments{
\item{file}{Fasta file}

\item{subset}{Vector with (exact) names of the sequences to be retrieved.}

\item{out}{Path to output file. If absent, the '.mod' is added to the input file's name (path is thus conserved).}
}
\value{
Writes a fasta file with the selected sequences.
}
\description{
This function loads a fasta file, selects a subset on sequences based on the sequence's names and writes a new fsata file.
}
\examples{
subset.fasta(file="http://greengenes.lbl.gov/Data/JD_Tutorial/UnAligSeq24606.txt",subset=c("24.6jsd1.Tut","24.6jsd2.Tut ","24.6jsd3.Tut "),out="out.fasta")
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{FastaUtils}

