% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaUtils.R
\name{fastaList}
\alias{fastaList}
\title{Lists contigs of a (multi) fasta file and
counts contig sizes}
\usage{
fastaList(fasta)
}
\arguments{
\item{fasta}{path of the fasta file}
}
\value{
number of contigs and contigs sizes
}
\description{
Lists contigs of a (multi) fasta file and
counts contig sizes
}
