% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastanalyze.R
\name{fastanalyze}
\alias{fastanalyze}
\title{fastanalyze() gives information about a fasta file}
\usage{
fastanalyze(fasta_file, metrics = TRUE, contigs = TRUE, plot = TRUE,
  out = "")
}
\arguments{
\item{fasta_file}{path of the fasta file to analyze}

\item{metrics}{boolean, indicates whether the metrics should be returned.
default: TRUE}

\item{contigs}{boolean, indicates whether each contig name and size be
returned. default: TRUE}

\item{plot}{boolean, indicates whether a plot should be generated
returned. default: TRUE}
}
\value{
metrics contigs and plot if requested.
}
\description{
It counts the number of contigs, their size and the total size.
  It also calculates the N50 / L50, the N90 / L90 and can print a synthetic
  plot suming up all these results nicely.
}
